+++ oci/database_db_system_resource_test.go
@@ -289,6 +289,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_basi
 					data_storage_size_in_gb = "256"
 					license_model = "LICENSE_INCLUDED"
 					node_count = "1"
+					fault_domains = ["FAULT-DOMAIN-1"]
 					db_home {
 						db_version = "12.1.0.2"
 						database {
@@ -316,6 +317,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_basi
 					resource.TestCheckResourceAttr(s.ResourceName, "data_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr(s.ResourceName, "license_model", "LICENSE_INCLUDED"),
 					resource.TestCheckResourceAttr(s.ResourceName, "node_count", "1"),
+					resource.TestCheckResourceAttr(s.ResourceName, "fault_domains.#", "1"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.db_version", "12.1.0.2"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.display_name", ""),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.database.0.admin_password", "BEstrO0ng_#11"),
@@ -359,6 +361,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allB
 					data_storage_size_in_gb = "256"
 					license_model = "LICENSE_INCLUDED"
 					node_count = "1"
+					fault_domains = ["FAULT-DOMAIN-1"]
 					db_home {
 						db_version = "12.1.0.2"
 						display_name = "-tf-db-home"
@@ -434,6 +437,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allB
 					resource.TestCheckResourceAttr(s.ResourceName, "data_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr(s.ResourceName, "license_model", "LICENSE_INCLUDED"),
 					resource.TestCheckResourceAttr(s.ResourceName, "node_count", "1"),
+					resource.TestCheckResourceAttr(s.ResourceName, "fault_domains.#", "1"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.db_version", "12.1.0.2"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.display_name", "-tf-db-home"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.database.0.admin_password", "BEstrO0ng_#11"),
@@ -479,6 +483,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allB
 
 					resource.TestCheckResourceAttrSet("data.oci_database_db_systems.t", "db_systems.0.data_storage_size_in_gb"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_systems.t", "db_systems.0.node_count"),
+					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.fault_domains.#", "1"),
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.state", string(database.DbSystemLifecycleStateAvailable)),
 
 					// DB Systems nested DB Home fields are not supported on the data source, so tests like below wont work if/until fetching the sub resource is implemented
@@ -556,6 +561,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allB
 					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.id"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.state"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.time_created"),
+					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.fault_domain"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.vnic_id"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.backup_vnic_id"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.software_storage_size_in_gb"),
@@ -567,6 +573,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allB
 					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "id"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "state"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "time_created"),
+					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "fault_domain"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "vnic_id"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "backup_vnic_id"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "software_storage_size_in_gb"),
@@ -604,6 +611,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					data_storage_size_in_gb = "256"
 					license_model = "LICENSE_INCLUDED"
 					node_count = "1"
+					fault_domains = ["FAULT-DOMAIN-1"]
 					db_home {
 						db_version = "12.1.0.2"
 						display_name = "-tf-db-home"
@@ -682,6 +690,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					resource.TestCheckResourceAttr(s.ResourceName, "data_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr(s.ResourceName, "data_storage_percentage", "80"),
 					resource.TestCheckResourceAttr(s.ResourceName, "node_count", "1"),
+					resource.TestCheckResourceAttr(s.ResourceName, "fault_domains.#", "1"),
 					resource.TestCheckResourceAttr(s.ResourceName, "reco_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr(s.ResourceName, "listener_port", "1521"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.db_version", "12.1.0.2"),
@@ -718,6 +727,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.data_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.data_storage_percentage", "80"),
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.node_count", "1"),
+					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.fault_domains.#", "1"),
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.reco_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.listener_port", "1521"),
 					resource.TestCheckResourceAttr("data.oci_database_db_systems.t", "db_systems.0.defined_tags.example-tag-namespace-all.example-tag", "originalValue"),
@@ -812,6 +822,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.id"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.state"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.time_created"),
+					resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.fault_domain"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.vnic_id"), // believe this is null when using FAKEHOSTSERIAL header
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_nodes.t", "db_nodes.0.backup_vnic_id"),
 					resource.TestCheckResourceAttr("data.oci_database_db_nodes.t", "db_nodes.0.software_storage_size_in_gb", "200"),
@@ -823,6 +834,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "id"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "state"),
 					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "time_created"),
+					resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "fault_domain"),
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "vnic_id"), // believe this is null when using FAKEHOSTSERIAL header
 					//resource.TestCheckResourceAttrSet("data.oci_database_db_node.t", "backup_vnic_id"),
 					resource.TestCheckResourceAttr("data.oci_database_db_node.t", "software_storage_size_in_gb", "200"),
@@ -851,6 +863,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					data_storage_size_in_gb = "512"
 					license_model = "LICENSE_INCLUDED"
 					node_count = "1"
+					fault_domains = ["FAULT-DOMAIN-1"]
 					db_home {
 						db_version = "12.1.0.2"
 						display_name = "-tf-db-home"
@@ -929,6 +942,7 @@ func (s *ResourceDatabaseDBSystemTestSuite) TestAccResourceDatabaseDBSystem_allV
 					resource.TestCheckResourceAttr(s.ResourceName, "data_storage_size_in_gb", "512"),
 					resource.TestCheckResourceAttr(s.ResourceName, "data_storage_percentage", "80"),
 					resource.TestCheckResourceAttr(s.ResourceName, "node_count", "1"),
+					resource.TestCheckResourceAttr(s.ResourceName, "fault_domains.#", "1"),
 					resource.TestCheckResourceAttr(s.ResourceName, "reco_storage_size_in_gb", "256"),
 					resource.TestCheckResourceAttr(s.ResourceName, "listener_port", "1521"),
 					resource.TestCheckResourceAttr(s.ResourceName, "db_home.0.db_version", "12.1.0.2"),